nvim-lua: visual select + f2 to run treesitter tests
.nvim-commands.lua is a lua file containing commands for neovim that add
keybindings for this project to the running instance. It has only a
single dependency on toggleterm (not including tree-sitter cli) which is
used to improve the terminal workflow in neovim.

Source with `:luafile %%`

After that press the F1 key to run the entire test suite.

To run a single test, visual select the test name in the corpus text
file and press F2.

Press F3 to run the py.test test suite for roadmapgen.