Make array_expression less greedy
I've been dealing with a really greedy array_expression for a while now.
TS always complains it conflicts with this and that and I usually
attempted to work around it. But with dealing with the comma problem,
that no longer worked and I had to attack it head on. Thankfully, TS
`-D` made this a little less painful than I had imagined.