Parse tuples/destructuring_assignment_for.txt
At 95% parser complete for implementing the Jakt samples, This test made
me discover that commas in function arguments are completely optional!
In fact, a lot of places in Jakt they are optional.

https://github.com/SerenityOS/jakt/pull/988#issuecomment-1198796846

This threw me for a bit of a loop because i've been assuming commas were
required in a lot of places where they are not.

This commit is very large, but implementing the fix took two days broke
around 150+ tests.

With that said, we are still not completely compliant: the
binary_expression is not 100% accurate (orderwise, is_expression is way
off) and array_literal needs a custom parser because it has trouble with
tuples and call_expressions. But at least there are no errors!