Add original/logic.txt
Pain was had trying to get tree sitter to parse logical and and generic
call expressions. After much poking and hoping, I managed to get the
write incantation in grammar.js to get it working without errors. Almost
dropped into the c parser for generic_type.

The key to getting it working was moving "and" and "or" to
binary_expression and removing "logical_expression" altogether.